No longer manually start the getty@* services.
There is systemd-logind already taking care of this at runtime when the
user switches to the corresponding VT. Furthermore systemd-logind is
smarter since it will avoid running it on an already used VT (such as vt1
which is nowadays used by gdm3). This also gets rid of the undesirable
boot delay that was likely introduced by one of the above conflicts.